pyproject.toml
[project]
name = "downify"
version = "0.1.0"
description = "Telegram bot that resolves Spotify links to release metadata and downloads matching legal audio from another provider."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.6",
]

[project.scripts]
downify-bot = "downify.bot:main"

[tool.ruff]
line-length = 100
target-version = "py311"

